Adds a new builtin step - filter_rows. Moves builtin_steps and tests to new files
