+++ phaser/__init__.py
@@ -30,7 +30,8 @@ from phaser.pipeline import Pipeline
 from phaser.constants import PHASER_ROW_NUM, ON_ERROR_WARN, ON_ERROR_COLLECT, ON_ERROR_DROP_ROW, ON_ERROR_STOP_NOW
 from phaser.exceptions import PhaserError, DataErrorException, DataException, DropRowException, WarningException
 from phaser.phase import Phase, ReshapePhase
-from phaser.steps import row_step, batch_step, dataframe_step, context_step, check_unique, sort_by
+from phaser.steps import row_step, batch_step, dataframe_step, context_step
+from phaser.builtin_steps import check_unique, sort_by, filter_rows
 from phaser.column import Column, IntColumn, DateColumn, DateTimeColumn, FloatColumn
 from phaser.io import read_csv, save_csv
